update Nav Bar & rename components for clarity
