--- .vscode/.gitignore
@@ -1,1 +1,0 @@
-settings.json
