+++ pages/index.js
@@ -1,9 +1,8 @@
 import styles from './index.module.scss';
 
-import Footer from '../components/Landing Page/Footer/footer';
-import { Nav } from '../components/Landing Page/Nav/nav';
+import Footer from '/components/Footer/footer';
+import { Nav } from '/components/Nav/nav';
 
-import Image from 'next/image';
 import Link from 'next/link';
 import Head from 'next/head';
